feat: add ability to embed images
default behavior is now to embed images, this helps resolve an issue where when md file is not in the same folder as output, image linking is broken. Safer/more consistent to just embed images understanding that it is rare for an image to be used in multiple markdowns. so the occasional bloat is acceptable